requirements.txt
pytest
selenium
pyyaml
webdriver-manager